[Fixes #74554302] - NPE Play Store okHttp
The null pointer is in the okhttp code; which we cannot edit.  We did, however, fork the Facebook sdk so I can make changes in there, so I added a try/catch to the disconnectQuietly method which is on the path of the NPE.  This will prevent the force close for this particular ticket.

QA: Not sure how to tell you to test this, probably the usual of just accept the ticket and watch for force closes.