[finishes #84779466] Change facebook api to v2
I pulled in the very latest code from facebook, version 3.22. I had to edit some code make this work on android api 8. The code had been built to work on api 9. These edits should not effect our app.

QA: Confirm that you are able to create an account through facebook. Confirm that you are able to login to an account with facebook.